# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Anime Multiverse Arena API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: getPlayers
      tags: [players]
      summary: List all players
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [level, gold, pvpRating, wins, createdAt]
            default: level
      responses:
        "200":
          description: Players list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayersPage"
  /players/{discordId}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player details
      parameters:
        - name: discordId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDetail"
  /players/{discordId}/ban:
    post:
      operationId: banPlayer
      tags: [players]
      summary: Ban or unban a player
      parameters:
        - name: discordId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanPlayerRequest"
      responses:
        "200":
          description: Ban status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /players/{discordId}/modify:
    post:
      operationId: modifyPlayer
      tags: [players]
      summary: Modify player resources (gold, gems, xp)
      parameters:
        - name: discordId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyPlayerRequest"
      responses:
        "200":
          description: Player modified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /players/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [players]
      summary: Get top players leaderboard
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [level, pvpRating, wins, totalDamageDealt]
            default: level
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
  /market/listings:
    get:
      operationId: getMarketListings
      tags: [market]
      summary: List market listings
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Market listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketPage"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    PlayersPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PlayerSummary"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, totalPages]
    PlayerSummary:
      type: object
      properties:
        id:
          type: integer
        discordId:
          type: string
        username:
          type: string
        level:
          type: integer
        gold:
          type: integer
        gems:
          type: integer
        pvpRating:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        isBanned:
          type: boolean
        currentWorld:
          type: string
        createdAt:
          type: string
      required: [id, discordId, username, level, gold, gems, pvpRating, wins, losses, isBanned, currentWorld, createdAt]
    PlayerDetail:
      type: object
      properties:
        player:
          $ref: "#/components/schemas/PlayerSummary"
        characterCount:
          type: integer
        totalDamageDealt:
          type: integer
        maxAbyssFloor:
          type: integer
        stamina:
          type: integer
        maxStamina:
          type: integer
        currentTitle:
          type: string
      required: [player, characterCount, totalDamageDealt, maxAbyssFloor, stamina, maxStamina]
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        discordId:
          type: string
        username:
          type: string
        value:
          type: integer
        level:
          type: integer
      required: [rank, discordId, username, value, level]
    BanPlayerRequest:
      type: object
      properties:
        ban:
          type: boolean
        reason:
          type: string
      required: [ban]
    ModifyPlayerRequest:
      type: object
      properties:
        gold:
          type: integer
        gems:
          type: integer
        xp:
          type: integer
        reason:
          type: string
...
    MarketPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MarketListingSummary"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, totalPages]
    MarketListingSummary:
      type: object
      properties:
        id:
          type: integer
        sellerUsername:
          type: string
        itemName:
          type: string
        itemRarity:
          type: string
        itemType:
          type: string
        listingType:
          type: string
        price:
          type: integer
        quantity:
          type: integer
        status:
          type: string
        createdAt:
          type: string
      required: [id, sellerUsername, itemName, itemRarity, itemType, listingType, price, quantity, status, createdAt]
...